+++ src/containers/vector.h
@@ -5,6 +5,9 @@
 #include "base/common.h"
 #include "memory/common.h"
 
+// External includes
+#include <algorithm>
+
 namespace Donut
 {
 	// Foward declaration
@@ -14,8 +17,8 @@ namespace Donut
 	class TVector
 	{
 		// Internal values
-		const uint MINIMAL_RESIZE_CAPACITY = 10;
-		const uint RESIZE_FACTOR = 2;
+		const uint64_t MINIMAL_RESIZE_CAPACITY = 10;
+		const uint64_t RESIZE_FACTOR = 2;
 		// Interval types
 		typedef T* iterator;
 
@@ -29,7 +32,7 @@ namespace Donut
 
 		}
 
-		TVector(TAllocator& _alloc, uint _reservedSize)
+		TVector(TAllocator& _alloc, uint64_t _reservedSize)
 		: m_alloc(_alloc)
 		, m_size(0)
 		, m_capacity(0)
@@ -49,10 +52,10 @@ namespace Donut
 			}
 		}
 
-		inline uint size() {return m_size;}
-		inline uint capacity() {return m_capacity;}
+		inline uint64_t size() {return m_size;}
+		inline uint64_t capacity() {return m_capacity;}
 
-		void resize(uint _size)
+		void resize(uint64_t _size)
 		{
 			if(_size ==  0)
 			{
@@ -112,7 +115,7 @@ namespace Donut
 			}
 		}
 
-		void reserve(uint _space)
+		void reserve(uint64_t _space)
 		{
 			ASSERT_NO_RELEASE(_space != 0);
 			memory_block mem = m_alloc.allocate(sizeof(T) * (m_capacity + _space));
@@ -153,8 +156,8 @@ namespace Donut
 
 	protected:
 		T* m_data;
-		uint m_size;
-		uint m_capacity;
+		uint64_t m_size;
+		uint64_t m_capacity;
 
 		TAllocator& m_alloc;
 	};
